fixup! feat: move rss3 web3 tab into rss3 plugin
