+++ packages/mask/shared-ui/locales/qya-AA.json
@@ -1012,8 +1012,5 @@
     "plugin_profile_address_name_dns": "crwdns10677:0crwdne10677:0",
     "plugin_profile_address_name_rns": "crwdns10679:0crwdne10679:0",
     "plugin_profile_address_name_address": "crwdns10681:0crwdne10681:0",
-    "plugin_profile_rss3_inactive_project": "crwdns10683:0crwdne10683:0",
-    "plugin_profile_rss3_no_activity_time": "crwdns10685:0crwdne10685:0",
-    "plugin_profile_rss3_attended": "crwdns10687:0crwdne10687:0",
     "plugin_vcent_last_offer_at": "crwdns10295:0crwdne10295:0"
 }
