+++ packages/mask/src/plugins/Profile/services.ts
@@ -1,11 +1,5 @@
 import urlcat from 'urlcat'
-import { fetchJSON } from '@masknet/plugin-infra'
-import { AssetType, GeneralAsset, RSS3Profile } from './types'
-
-interface Response {
-    status: boolean
-    assets: GeneralAsset[]
-}
+import { fetchJSON } from '../../extension/background-script/HelperService'
 
 interface NameInfo {
     rnsName: string
@@ -13,38 +7,9 @@ interface NameInfo {
     address: string
 }
 
-interface RSS3Info {
-    profile: RSS3Profile
-}
-
-export function getDonations(address: string) {
-    const url = urlcat('https://hub.pass3.me/assets/list', {
-        personaID: address,
-        type: AssetType.GitcoinDonation,
-    })
-
-    return fetchJSON<Response>(url)
-}
-
-export function getFootprints(address: string): Promise<Response> {
-    const url = urlcat('https://hub.pass3.me/assets/list', {
-        personaID: address,
-        type: AssetType.POAP,
-    })
-
-    return fetchJSON<Response>(url)
-}
-
 export async function getRSS3AddressById(id: string) {
     if (!id) return ''
     const url = urlcat('https://rss3.domains/name/:id', { id })
     const rsp = await fetchJSON<NameInfo>(url)
     return rsp.address
-}
-
-export async function getRSS3ProfileByAddress(address: string) {
-    if (!address) return
-    const url = urlcat('https://hub.pass3.me/:address', { address })
-    const rsp = await fetchJSON<RSS3Info>(url)
-    return rsp?.profile
 }
