--- packages/plugin-infra/src/apis/index.ts
@@ -1,1 +1,0 @@
-export * from './network'
